+++ NEWS.md
@@ -1,8 +1,9 @@
 # lcars 0.2.0
 
 * Added support for a minimal LCARS theme Shiny UI.
-* New functions: `lcarsPage`, `lcarsButton`, `lcarsCheckbox`, `lcarsToggle`, `buttonColumn`, `lcarsRadio`, `lcarsHeader`, `lcarsBox` and `lcarsApp`.
+* New functions: `lcarsPage`, `lcarsButton`, `lcarsCheckbox`, `lcarsToggle`, `buttonColumn`, `lcarsRadio`, `lcarsHeader` and other header functions, `lcarsBox`, `lcarsWell` and `lcarsApp`.
 * Added named LCARS colors from different palettes; see `lcarsColors`.
+* Added readme and vignette content.
 
 # lcars 0.1.0